chore: seperate out project componenet for reusability
