+++ web/core/components/project/index.ts
@@ -18,4 +18,6 @@ export * from "./member-list-item";
 export * from "./project-settings-member-defaults";
 export * from "./send-project-invitation-modal";
 export * from "./confirm-project-member-remove";
+export * from "./multi-select-modal";
+export * from "./search-projects";
 export * from "@/plane-web/components/projects/create/root";
